+++ src/main/java/dev/orne/log/manager/Level.java
@@ -1,0 +1,235 @@
+package dev.orne.log.manager;
+
+import java.io.Serializable;
+import java.util.Objects;
+
+import org.apache.commons.lang3.builder.CompareToBuilder;
+import org.apiguardian.api.API;
+import org.jspecify.annotations.Nullable;
+
+/**
+ * Base, logging system agnostic, log level.
+ * 
+ * @author <a href="https://github.com/ihernaez">(w) Iker Hernaez</a>
+ * @version 1.0, 2025-12
+ * @since 1.0
+ */
+@API(status = API.Status.STABLE, since = "1.0.0")
+public class Level
+implements Serializable, Comparable<Level> {
+
+    /** The serial version UID. */
+    private static final long serialVersionUID = 1L;
+
+    /** The name of the level. */
+    private final String name;
+    /** The numeric value of the level. */
+    private final int value;
+
+    /**
+     * Creates a new instance.
+     * 
+     * @param builder The builder with the instance state
+     */
+    protected Level(
+            final BuilderImpl builder) {
+        super();
+        this.name = Objects.requireNonNull(builder.name, "The name must be set");
+        this.value = Objects.requireNonNull(builder.value, "The value must be set");
+    }
+
+    /**
+     * Creates an empty builder.
+     * 
+     * @return The created builder
+     */
+    public static Builder builder() {
+        return new BuilderImpl();
+    }
+
+    /**
+     * Creates a builder with an initial state copied from the specified
+     * instance.
+     * 
+     * @param copy The instance to copy
+     * @return The created builder
+     */
+    public static Builder copyOf(
+            final Level copy) {
+        return new BuilderImpl(copy);
+    }
+
+    /**
+     * Returns the name of the level
+     * 
+     * @return The name of the level
+     */
+    public String getName() {
+        return this.name;
+    }
+
+    /**
+     * Returns the numeric value of the level.
+     * 
+     * @return The numeric value of the level
+     */
+    public int getValue() {
+        return this.value;
+    }
+
+    /**
+     * Creates a builder with an initial state copied from this instance.
+     * 
+     * @return The created builder
+     */
+    public Builder copy() {
+        return new BuilderImpl(this);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public int hashCode() {
+        return Objects.hash(
+                this.name,
+                this.value);
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public boolean equals(
+            final Object obj) {
+        if (this == obj) {
+            return true;
+        }
+        if (obj == null) {
+            return false;
+        }
+        if (getClass() != obj.getClass()) {
+            return false;
+        }
+        final Level other = (Level) obj;
+        return Objects.equals(name, other.name) && value == other.value;
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public int compareTo(final Level other) {
+        return new CompareToBuilder()
+                .append(this.value, other.getValue())
+                .toComparison();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public String toString() {
+        return this.name;
+    }
+
+    /**
+     * Builder interface for {@link Level} instances.
+     * 
+     * @author <a href="https://github.com/ihernaez">(w) Iker Hernaez</a>
+     * @version 1.0, 2025-12
+     * @since 1.0
+     */
+    @API(status = API.Status.STABLE, since = "1.0.0")
+    public interface Builder {
+
+        /**
+         * Sets the name of the level.
+         * 
+         * @param name The name of the level
+         * @return This builder, for method chaining
+         */
+        Builder withName(
+                String name);
+
+        /**
+         * Sets the numeric value of the level.
+         * 
+         * @param value The numeric value of the level
+         * @return This builder, for method chaining
+         */
+        Builder withValue(
+                int value);
+
+        /**
+         * Builds the level instance.
+         * 
+         * @return The built level instance
+         */
+        Level build();
+    }
+
+    /**
+     * Builder implementation for {@link Level} instances.
+     * 
+     * @author <a href="https://github.com/ihernaez">(w) Iker Hernaez</a>
+     * @version 1.0, 2025-12
+     * @since 1.0
+     */
+    @API(status = API.Status.STABLE, since = "1.0.0")
+    static class BuilderImpl implements Builder {
+
+        /** The name of the level. */
+        private @Nullable String name;
+        /** The numeric value of the level. */
+        private @Nullable Integer value;
+
+        /**
+         * Empty constructor.
+         */
+        public BuilderImpl() {
+            super();
+        }
+
+        /**
+         * Copy constructor.
+         * 
+         * @param copy The instance to copy
+         */
+        public BuilderImpl(
+                final Level copy) {
+            super();
+            Objects.requireNonNull(copy, "Instance to copy cannot be null");
+            this.name = copy.getName();
+            this.value = copy.getValue();
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public BuilderImpl withName(
+                final String name) {
+            this.name = name;
+            return this;
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public BuilderImpl withValue(
+                final int value) {
+            this.value = value;
+            return this;
+        }
+
+        /**
+         * {@inheritDoc}
+         */
+        @Override
+        public Level build() {
+            return new Level(this);
+        }
+    }
+}