+++ CHANGELOG.md
@@ -3,3 +3,8 @@
 ## 1.0.0 - *Unreleased*
 
 **Initial Release**
+
+### Added
+
+- Log management API
+    - `Level` immutable bean for logging system agnostic log levels